+++ adapter/spark/src/test/java/com/asml/apa/wta/spark/BaseSparkJobIntegrationTest.java
@@ -3,6 +3,7 @@ package com.asml.apa.wta.spark;
 import com.asml.apa.wta.core.config.RuntimeConfig;
 import com.asml.apa.wta.core.model.enums.Domain;
 import com.asml.apa.wta.spark.datasource.SparkDataSource;
+import com.asml.apa.wta.spark.streams.MetricStreamingEngine;
 import java.util.Arrays;
 import org.apache.spark.SparkConf;
 import org.apache.spark.api.java.JavaRDD;
@@ -21,6 +22,8 @@ public class BaseSparkJobIntegrationTest {
 
   RuntimeConfig fakeConfig;
 
+  MetricStreamingEngine fakeMetricStreamingEngine;
+
   @BeforeEach
   void setupBaseIntegrationTest() {
     fakeConfig = RuntimeConfig.builder()
@@ -37,6 +40,8 @@ public class BaseSparkJobIntegrationTest {
     SparkConf conf = new SparkConf().setAppName("SparkTestRunner").setMaster("local");
     spark = SparkSession.builder().config(conf).getOrCreate();
     spark.sparkContext().setLogLevel("ERROR");
+
+    fakeMetricStreamingEngine = new MetricStreamingEngine();
 
     sut = new SparkDataSource(spark.sparkContext(), fakeConfig);
     String resourcePath = "src/test/resources/wordcount.txt";