+++ adapter/spark/src/test/java/com/asml/apa/wta/spark/listener/JobLevelListenerTest.java
@@ -31,7 +31,7 @@ class JobLevelListenerTest extends BaseLevelListenerTest {
     Workflow fakeJobListenerWorkflow = fakeJobListener.getProcessedObjects().get(0);
     assertThat(fakeJobListenerWorkflow.getId()).isEqualTo(560);
     assertThat(fakeJobListenerWorkflow.getTsSubmit()).isEqualTo(40L);
-    assertThat(fakeJobListenerWorkflow.getScheduler()).isEqualTo("DAGScheduler");
+    assertThat(fakeJobListenerWorkflow.getScheduler()).isEqualTo("FIFO");
     assertThat(fakeJobListenerWorkflow.getApplicationName()).isEqualTo("testApp");
 
     assertThat(fakeJobListenerWorkflow.getNfrs()).isEmpty();